Implemented mock classes and the ability to switch between using mock data and actual database data using appsettings.json. Now the REST service funcitonalities are fully testable.

+++ RESTservice-API/Models/Asset.cs
@@ -5,9 +5,8 @@ namespace RESTservice_API.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public int FloorMapId { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         public bool Active { get; set; }
     }
 }
-
